Arduino Addition
This update creates an Arduino folder for all Arduino code.
Currently, this includes a version of the cabinet software which can
activate and deactivate the backlight using either the RFID wand or an
XBee signal.

Additionally, simple test sketches are included in their own folder.
These sketches are for testing individual hardware pieces, both to
ensure they are working properly and as examples to the user of how to
interface with them.