Concept of "Game" hw+subprocess, tons of script cleanup.
The Game hardware object will control the state of the game and fire
events. Instead of relying on Countdown.Expired for the lose scenario,
for example. We will instead listen for events on the Game itself. Next
update, in the UI we will have direct control over the game state.